+++ src/components/chat/answer/feedbackButtons.tsx
@@ -11,8 +11,9 @@ import { useAIState, useActions } from 'ai/rsc';
 
 import AnimatedButton from '@/components/ui/animatedButton';
 import { AI } from '@/lib/actions';
-import { GTagEvents, gtagEvent } from '@/lib/gtag';
+import { useGtagEvent } from '@/lib/hooks/useGtagEvent';
 import { ChatHistory, Feedback } from '@/models/chat';
+import { GTagEvents } from '@/models/gtag';
 
 export type FeedbackLoadingStates = '' | Feedback;
 
@@ -22,6 +23,7 @@ const FeedbackButtons: React.FC<FeedbackButtonsProps> = ({}) => {
   const [aiState, setAIState] = useAIState<typeof AI>();
   const { id: chatId, reaction: feedback } = aiState;
   const { submitFeedback } = useActions<typeof AI>();
+  const gtagEvent = useGtagEvent();
 
   const onFeedback = async (newFeedback: Feedback) => {
     const newAiState: ChatHistory = {
